celestin's comment + refactor exceptions
